requirements.txt
torch>=1.10
torchvision>=0.11
torchtext>=0.11
labml>=0.4.133
labml-helpers>=0.4.84
numpy>=1.19
matplotlib>=3.0.3
einops>=0.3.0
gym[atari]
opencv-python
Pillow>=6.2.1